pyproject.toml
[project]
name = "receipt-review-software"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pyside6>=6.11.0",
    "python-dotenv>=1.2.2",
    "requests>=2.33.1",
]

[dependency-groups]
dev = [
    "pyinstaller>=6.19.0",
    "pytest>=9.0.3",
    "requests-mock>=1.12.1",
]

[tool.pytest.ini_options]
pythonpath = ["src"]
